Move sky parameters into frame graph
